`BrowserWrapper`의 일부 함수를 demo 용으로 분리
