+++ demo/src/redux/Viewer.reducer.js
@@ -4,10 +4,10 @@ import path, { initialState } from './Viewer.path';
 import { ImmutableObjectBuilder } from '../../../src/util/ImmutabilityHelper';
 import { updateObject } from '../../../src/util/Util';
 import createReducer from '../../../src/util/Reducer';
-import * as BrowserWrapper from '../../../src/util/BrowserWrapper';
+import { scrollTop, scrollHeight, screenHeight } from '../utils/BrowserWrapper';
 
-const isScrolledToTop = () => BrowserWrapper.scrollTop() <= 100;
-const isScrolledToBottom = () => BrowserWrapper.scrollTop() >= BrowserWrapper.scrollHeight() - BrowserWrapper.screenHeight() - 100;
+const isScrolledToTop = () => scrollTop() <= 100;
+const isScrolledToBottom = () => scrollTop() >= scrollHeight() - screenHeight() - 100;
 
 const isEdgeOfScreen = () => (isScrolledToTop() || isScrolledToBottom());
